refacotr/LJJ(#112): date누르는 컴포넌트 분리
