+++ src/components/Blog/BlogList/PrivateBlogList.tsx
@@ -1,23 +1,14 @@
-import { Svg } from '@/components/Svg';
 import * as S from './style';
 import { BlogListEachHoneyCard } from './BlogListEachHoneyCard';
-import { useTheme } from 'styled-components';
 import { useMemo } from 'react';
-import { date } from '@/utils';
 import { GetPrivateBlogPaginationQuery } from '@/apis/blog/queries';
 import { Link } from 'react-router-dom';
 import useObserver from '@/hook/useObserver';
 import { useFilterPageStore } from '@/store/paginationStore/useFilterPageStore';
-import {
-  onChangeYearHandler,
-  onClickYearPrevAndNextHandler,
-  onFilterHandler,
-} from './utils';
+import SelectBlogDateSection from './SelectBlogDateSection';
 
 export default function PrivateBlogList({ id }: { id: string }) {
-  const theme = useTheme();
-
-  const { year, month, filter, setMonth } = useFilterPageStore();
+  const { year, month, filter } = useFilterPageStore();
 
   const getBlogFilter = useMemo(
     () => ({
@@ -38,49 +29,7 @@ export default function PrivateBlogList({ id }: { id: string }) {
 
   return (
     <S.BlogListWrapper>
-      <S.BlogListHeaderWrapper>
-        <div>
-          <S.BlogListFilterSelect onChange={onFilterHandler} value={filter}>
-            <option value="all">전체</option>
-            <option value="public">공개글</option>
-          </S.BlogListFilterSelect>
-          <S.BlogSelectYearButtonWrapper>
-            <button onClick={() => onClickYearPrevAndNextHandler('prev')}>
-              <Svg.PrevIcon color={theme.text.primary} />
-            </button>
-            <span>
-              <input
-                type="date"
-                id="selectYear"
-                onChange={onChangeYearHandler}
-                placeholder={year}
-              />
-              년 의 달콤한 이야기
-            </span>
-            <button onClick={() => onClickYearPrevAndNextHandler('next')}>
-              <Svg.NextIcon color={theme.text.primary} />
-            </button>
-          </S.BlogSelectYearButtonWrapper>
-        </div>
-        <div>
-          {Array.from({ length: 12 }, (_, i) => (
-            <S.BlogSelectMonthButton
-              key={`${i + 1}-month-filter`}
-              $isSelectedMonth={
-                month ===
-                date.getFormattingDate({ date: i + 1, formatType: 'MM' })
-              }
-              onClick={() =>
-                setMonth(
-                  date.getFormattingDate({ date: i + 1, formatType: 'MM' })
-                )
-              }
-            >
-              {i + 1} 월
-            </S.BlogSelectMonthButton>
-          ))}
-        </div>
-      </S.BlogListHeaderWrapper>
+      <SelectBlogDateSection />
       <S.BlogSelectMonthSpan>{1}월</S.BlogSelectMonthSpan>
       {getBlogInfo?.data?.pages[0].contents.length !== 0 ? (
         <S.BlogListPaginationWrapper>
